refactor(solana): isolate user decrypt validation
